Добавил метод для создания случайного элемента (в соотв. с шансом) + новый класс LevelController

+++ Assets/Scripts/Elements/Element.cs
@@ -4,5 +4,5 @@ public class Element
 {
     public string Name { get; set; }
     public FieldState[,] Matrix { get; set; }
-    public float SpawnChance { get; set; }
+    public int SpawnChance { get; set; }
 }
